Assertion Generation:
JS Programmers may find our formulae and diagrams difficult to
interpret. This commit adds a module which takes a formula and
generates a collection of qunit assertions which dynamically check if
the formula is satisfied in the current heap. These assertions are an
under-approximation to limit false positives.

Currently the module and its associated test-set is empty. This will change!