Add one of the example specs to sanitiseImg
Trying to get a feel for what these specs will look like. The one I've added is the one that has a proof sketch expanded in the DOM paper. I skipped the val predicate and did a simpler case of a single text node, this may be too inaccurate.

The list part of the specification, where I try to describe the shape of the attributes of the element node may pose issues or be in the incorrect format.